refactor: split program and programYear routes
